app/requirements3.txt
altgraph @ file:///AppleInternal/Library/BuildRoots/bd680bd8-563b-11ef-bf16-8a553ba56670/Library/Caches/com.apple.xbs/Sources/python3/altgraph-0.17.2-py2.py3-none-any.whl
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
asttokens==2.2.0
attrs==22.1.0
audiosegment==0.23.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==43.0.1
debugpy==1.6.4
decorator==4.4.2
defusedxml==0.7.1
Django==4.2.16
django-cors-headers==3.13.0
django-embed-video==1.4.8
django-extensions==3.2.3
django-resized==1.0.2
djangorestframework==3.15.2
dnspython==2.2.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
ffmpeg==1.4
ffmpeg-python==0.2.0
ffprobe==0.5
flutter==0.1
future==0.18.2
idna==3.4
imageio==2.32.0
imageio-ffmpeg==0.4.9
importlib-metadata==5.1.0
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
macholib @ file:///AppleInternal/Library/BuildRoots/bd680bd8-563b-11ef-bf16-8a553ba56670/Library/Caches/com.apple.xbs/Sources/python3/macholib-1.15.2-py2.py3-none-any.whl
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
mongo==0.2.0
moviepy==1.0.3
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.24.1
oauthlib==3.2.2
packaging==21.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.5.4
proglog==0.1.10
prometheus-client==0.15.0
prompt-toolkit==3.0.33
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
PyAudio==0.2.14
pycparser==2.21
pydub==0.25.1
Pygments==2.13.0
PyJWT==2.9.0
pymongo==4.3.3
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
python3-openid==3.2.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
requests-oauthlib==2.0.0
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
terminado==0.17.0
tinycss2==1.2.1
tornado==6.2
tqdm==4.66.1
traitlets==5.6.0
typing_extensions==4.12.2
urllib3==1.26.12
virtualenv==20.16.7
wcwidth==0.2.5
webencodings==0.5.1
webrtcvad==2.0.10
websocket-client==1.4.2
whitenoise==6.2.0
widgetsnbextension==4.0.3
zipp==3.11.0
